sentiment-trader/collector/requirements.txt
tweepy
praw
python-dotenv
pandas
schedule
requests
